6. Refactor pages folder content for more compliance with FSD - pages should be more declarative, mostly consist of widgets.
