+++ src/pages/user_todos/user_todos.tsx
@@ -1,8 +1,5 @@
-import styles from './user_todos.module.scss';
 import { useNavigate, useParams } from 'react-router-dom';
-import { Todo } from '../../entities/todo/ui/Todo.tsx';
-import { useGetTodosByUserIdQuery } from '../../entities/todo/api/todosApi.ts.ts';
-import { ItemList } from '../../shared/ui/ItemList/ItemList.tsx';
+import { TodosList } from '../../widgets/TodosList/TodosList.tsx';
 
 export const UserPageTodos = () => {
   const { id: userId } = useParams();
@@ -10,23 +7,5 @@ export const UserPageTodos = () => {
 
   if (!userId) navigate('/posts');
 
-  const { data: userTodos = [], isLoading } = useGetTodosByUserIdQuery(userId!);
-
-  return isLoading ? (
-    <h3>Загрузка...</h3>
-  ) : (
-    <div className={styles.todosListWrapper}>
-      <ItemList
-        items={userTodos}
-        keyExtractor={(todo) => todo.id}
-        container={'ul'}
-        extClassName={styles.todosList}
-        renderItem={(todo) => (
-          <li className={styles.todoItem}>
-            <Todo todo={todo} />
-          </li>
-        )}
-      />
-    </div>
-  );
+  return <TodosList userId={userId} />;
 };
