+++ src/widgets/AlbumsList/AlbumsList.module.scss
